#798 and #838?
