+++ src/box/CMakeLists.txt
@@ -47,6 +47,7 @@ add_library(box
     request.cc
     txn.cc
     box.cc
+    box_stat.c
     user_def.cc
     user.cc
     authentication.cc
