+++ test/box/misc.result
@@ -135,12 +135,13 @@ end;
 t;
 ---
 - - DELETE
-  - EVAL
   - SELECT
-  - REPLACE
   - INSERT
-  - AUTH
+  - EVAL
+  - EXCEPTION
   - CALL
+  - REPLACE
+  - AUTH
   - UPDATE
   - total
   - rps
